+++ CHANGES.rst
@@ -5,6 +5,10 @@ Changelog
 1.2.57 (unreleased)
 -------------------
 
+- Refactor : Move imio.smartweb.vocabulary.ContactBlocks from imio.smartweb.core to imio.smartweb.common 
+  (usefull for linked contact in auth sources)
+  [boulch]
+
 - SUP-53555 : Use schema.TextLine (with validator) instead of schema.Int for zipcode field to avoid an awkward space 
   (ex : 5 300 when editing the field) + upgrade steps
   [boulch]